pyproject.toml
[project]
name = "oquno-gohan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.3",
    "requests-oauthlib>=2.0.0",
    "tweepy>=4.15.0",
]
